[MERGE] permut block integration

+++ includes/disasm.h
@@ -57,7 +57,7 @@
 struct operands
 {
 	void		*addr;		/* instruction address   */
-	size_t		length;		/* instruction length    */
+	uint8_t		length;		/* instruction length    */
 	uint32_t	src;		/* source operand        */
 	uint32_t	dst;		/* destination operand   */
 };
@@ -67,18 +67,18 @@ struct operands
 */
 struct control_flow
 {
-	void            *addr;        /* instruction address             */
-	size_t          length;       /* instruction length              */
-	void            *value_addr;  /* value address in instruction    */
-	int8_t          value_length; /* value size in byte              */
-	int32_t         value;        /* immediate value in instruction  */
-	void            *label_addr;  /* label address                   */
+	void		*addr;        /* instruction address             */
+	uint8_t		length;       /* instruction length              */
+	void		*value_addr;  /* value address in instruction    */
+	uint8_t		value_length; /* value size in byte              */
+	int32_t		value;        /* immediate value in instruction  */
+	void		*label_addr;  /* label address                   */
 };
 
 /*
 ** disassembly functions
 */
-size_t		disasm_length(const void *code, size_t codelen);
+uint8_t		disasm_length(const void *code, size_t codelen);
 size_t		disasm_operands(const void *code, size_t codelen,
 			struct operands *buf, size_t buflen);
 size_t		disasm_jumps(struct control_flow *buf, size_t buflen,
